Split files a little bit
